+++ client/rpc-core/src/types/mod.rs
@@ -52,9 +52,6 @@ pub use self::sync::{
 	ChainStatus, EthProtocolInfo, PeerCount, PeerInfo, PeerNetworkInfo, PeerProtocolsInfo, Peers,
 	PipProtocolInfo, SyncInfo, SyncStatus, TransactionStats,
 };
-pub use self::transaction::{
-	LocalTransactionStatus, PendingTransaction, PendingTransactions, RichRawTransaction,
-	Transaction,
-};
+pub use self::transaction::{LocalTransactionStatus, RichRawTransaction, Transaction};
 pub use self::transaction_request::TransactionRequest;
 pub use self::work::Work;