+++ client/rpc-core/CHANGELOG.md
@@ -2,3 +2,4 @@
 
 ## Unreleased
 - Add `FilteredParams::address_in_bloom()` and `FilteredParams::topics_in_bloom()` functions to check the possible existance of Filter addresses or topics in a block.
+- Removed `PendingTransaction` and `PendingTransactions` types.